requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
marshmallow==3.20.1
marshmallow-sqlalchemy==0.30.0
SQLAlchemy==1.4.48
psycopg2-binary==2.9.6
python-dotenv==1.0.0
click==8.1.8
Jinja2==3.1.6
Werkzeug==2.3.7
itsdangerous==2.2.0
cloudinary==1.36.0
sendgrid==6.10.0
algoliasearch==3.0.0
gunicorn==23.0.0
